+++ BlazorApp3/Components/Pages/Users.razor
@@ -1,0 +1,42 @@
+﻿@page "/users"
+
+@using BlazorApp3.Model
+
+
+
+<h3>Users</h3>
+
+<div class="row">
+	<div class="col">
+		<h1 class="text-info">Add user</h1>
+	</div>
+</div>
+
+<div class="row">
+	<div class="col">
+		<EditForm Model="newUser" OnSubmit="addUser" FormName="AddUser">
+			<div class="form-group">
+				<label>Login:</label>
+				<InputText class="form-control" @bind-Value="newUser!.Login" />
+			</div>
+			<button type="submit">Save</button>
+		</EditForm>
+	</div>
+</div>
+
+
+
+@code {
+	ApplicationUser newUser;
+
+	protected override void OnInitialized() => newUser ??= new();
+
+	void addUser()
+	{
+		Console.WriteLine(newUser.Login);
+		UserStorage.Users.Add(newUser);
+	}
+
+
+
+}
