+++ BlazorApp3/Model/UserStorage.cs
@@ -4,7 +4,7 @@ namespace BlazorApp3.Model
 {
 	static public  class UserStorage
 	{
-		private static  List<ApplicationUser> _users;
+		private static  List<ApplicationUser> _users = new();
 		public static List<ApplicationUser>  Users { get { return _users; }  set { _users = value; } }
 		
 	}
